change4

+++ src/main/java/org/example/vet/service/DoctorService.java
@@ -2,9 +2,9 @@ package org.example.vet.service;
 
 import lombok.RequiredArgsConstructor;
 import org.example.vet.DTO.DoctorDTO;
-import org.example.vet.controllers.DoctorController;
 import org.example.vet.convertors.DoctorConvertor;
 import org.example.vet.entety.Doctor;
+import org.example.vet.exceptions.DoctorWithIdDoesNotExistException;
 import org.example.vet.repository.DoctorRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -20,38 +20,36 @@ public class DoctorService {
     @Autowired
     private final DoctorConvertor doctorConvertor;
 
-//    @Autowired
-//    public DoctorService(DoctorRepository doctorRepository, DoctorConvertor doctorConvertor) {
-//        this.doctorRepository = doctorRepository;
-//        this.doctorConvertor = doctorConvertor;
-//    }
-
     public List<DoctorDTO> getAll() {
         List<Doctor> doctors = doctorRepository.getAll();
-        return doctors.stream().map(doctorConvertor::convert).collect(Collectors.toList());
+        return doctors.stream().map(this::convertDoctor).collect(Collectors.toList());
     }
 
     public DoctorDTO findById(Long doctorId) {
         Doctor doctor = doctorRepository.findDoctorById(doctorId);
-        DoctorDTO doctorDTO = doctorConvertor.convert(doctor);
+        DoctorDTO doctorDTO = convertDoctor(doctor);
         return doctorDTO;
     }
 
     public void deleteById(Long doctorId) {
+        if (doctorRepository.findDoctorById(doctorId) == null) {
+            throw new DoctorWithIdDoesNotExistException("Doctor with id %s does not exist", doctorId);
+        }
         doctorRepository.deleteDoctorById(doctorId);
     }
 
     public void update(DoctorDTO doctorDTO, Long doctorId) {
-        doctorRepository.updateDoctor(doctorDTO, doctorId);
+        doctorRepository.updateDoctor(convertDTO(doctorDTO), doctorId);
     }
 
     public void insert(DoctorDTO doctorDTO) {
-        doctorRepository.insertDoctor(doctorDTO);
+        doctorRepository.insertDoctor(convertDTO(doctorDTO));
     }
 
-    public DoctorDTO convert(Doctor doctor)  {
+    public DoctorDTO convertDoctor(Doctor doctor)  {
         return doctorConvertor.convert(doctor);
     }
+    public Doctor convertDTO(DoctorDTO doctorDTO)  { return doctorConvertor.convert(doctorDTO); }
 }
